requirements.txt
flask==1.1.2
psycopg2-binary==2.9.5
jinja2<3.1.0
itsdangerous==2.0.1
werkzeug==2.0.3
